requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
idna==2.8
optionaldict==0.1.1
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
wechatpy==1.8.3
xmltodict==0.12.0
